fix(ui,api): fix handling errors (closes #1)
Solved handling errors by creating a function, `errorType`, that
receives a status code, and returns the error type based on that status
code.

This function was used in the `authRequest` function, which now returns
two extra properties, `error`, which is a boolean and `type`, which is
the error type.

Now, using that `type`, we can handle errors efficiently, for example:
- if a user is Unauthorized 401 (expired cookie, or signed out), will
  immediately redirect the user to `/signin`.
- if the user tried to access a file that's Not Found (404).